--- GPTEveningFeast/src/main/java/kosa/hdit5/evenapp/service/SigninService.java
@@ -1,8 +1,0 @@
-package kosa.hdit5.evenapp.service;
-
-import kosa.hdit5.evenapp.vo.UserVO;
-
-public interface SigninService {
-	
-	public UserVO selectOneUser(String userId);
-}
