+++ GPTEveningFeast/src/main/java/kosa/hdit5/evenapp/service/UserService.java
@@ -2,8 +2,9 @@ package kosa.hdit5.evenapp.service;
 
 import kosa.hdit5.evenapp.vo.UserVO;
 
-public interface SignupService {
+public interface UserService {
 	
+	public UserVO selectOneUser(String userId);
 	public boolean createUser(UserVO user);
 	public boolean validateUniqueUserId(String userId);
 }
