+++ GPTEveningFeast/src/main/java/kosa/hdit5/evenapp/service/UserServiceImpl.java
@@ -7,10 +7,22 @@ import kosa.hdit5.evenapp.mapper.UserMapper;
 import kosa.hdit5.evenapp.vo.UserVO;
 
 @Service
-public class SigninServiceImpl implements SigninService {
-
+public class UserServiceImpl implements UserService {
+	
 	@Autowired
 	private UserMapper mapper;
+
+	@Override
+	public boolean createUser(UserVO user) {
+		int result = mapper.insertUser(user);
+		return result == 1 ? true : false;
+	}
+
+	@Override
+	public boolean validateUniqueUserId(String userId) {
+		UserVO result = mapper.selectOneUser(userId);
+		return result == null ? true : false;
+	}
 	
 	@Override
 	public UserVO selectOneUser(String userId) {
